As we met problem to get the expected symbols at the beginning of the simulation, i have made a few changes to get it work.
1: from_mac happenned to have a value at time '0' which is not right since we only
have a value at Fb

2: Fb and Fs went to '1' at time '0', while there were no relevant data to prrocess
and produce irrelevant symbol at beginning of the simulation.

3: At Fs, we needed the bit at 4*Fb to output the related symbol, wehreas Fs = 4*Fb.
It is not possible to make this due to the delay time of input-output propagation.
So instead of firing Fs at mod 4000 it will be 4016, 16*4 = 64ns which represents the delay
time.
Also, At Fs, it is not possible to get instantly the symbol at the output,
this symbol will be available at Fs + 1.That why our counter as to be greater that 4016,
which actually means 8032 to write into the 'output_symbols.txt' file.

4: I reduced de clock time from 62.5ns to 4ns for 2 reasons :
	- Modelsim seems to round the value to 63ns and "period/2=31.5ns" to 32ns and affected
	  all the simulation time base.

	- Due to the delay time it will be necessary to increase the scale to reduce the delay time,
	for example if clk = 64ns, a 1 scale -> 64ns*1 = 64ns
	while if clk = 4ns a 1 scale -> 5ns, so the delay time is more manageable in the the later case.