This file contains the transmitter program which includes : - bit_to_symbol - symbol_to_chip - shaping
here we read the bits from "input_bits.txt" and only output the samples ich1
and qch1 because the others values (symbol,ich0 and qch0) have to be both read
and write which can't be done without a significant delay.

The results from the simulation are correct and predictable.

Now we have to edit the "tx.vhd" to make the transmitter program and also testing
file.That is the last task of our work before optimisations consideration.